refactor: code review 반영
