+++ src/main/java/com/nineteen/omp/store/controller/ServiceAreaController.java
@@ -1,20 +1,21 @@
 package com.nineteen.omp.store.controller;
 
 import com.nineteen.omp.global.dto.ResponseDto;
+import com.nineteen.omp.global.utils.PageableUtils;
 import com.nineteen.omp.store.controller.dto.ServiceAreaRequestDto;
 import com.nineteen.omp.store.controller.dto.ServiceAreaResponseDto;
 import com.nineteen.omp.store.domain.Area;
 import com.nineteen.omp.store.domain.Store;
-import com.nineteen.omp.store.service.AreaService;
 import com.nineteen.omp.store.service.ServiceAreaService;
-import com.nineteen.omp.store.service.StoreService;
 import com.nineteen.omp.store.service.dto.ServiceAreaCommand;
-import com.nineteen.omp.store.util.PageableUtils;
+import jakarta.validation.Valid;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort.Direction;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -31,17 +32,20 @@ import org.springframework.web.bind.annotation.RestController;
 public class ServiceAreaController {
 
   private final ServiceAreaService serviceAreaService;
-  private final AreaService areaService;
-  private final StoreService storeService;
+//  private final AreaService areaService;
+//  private final StoreService storeService;
 
   @PostMapping
   public ResponseEntity<ResponseDto<ServiceAreaResponseDto>> createServiceArea(
-      @RequestBody ServiceAreaRequestDto serviceAreaRequestDto) {
+      @Valid @RequestBody ServiceAreaRequestDto serviceAreaRequestDto) {
 
-    Area area = areaService.getArea(serviceAreaRequestDto.areaId());
-    Store store = storeService.getStore(serviceAreaRequestDto.storeId());
-    System.out.println("확인" + area + ",   " + store);
-    //command 이용
+    //통합시 응답 dto -> 엔티티 매핑 필요
+//    Area area = areaService.getArea(serviceAreaRequestDto.areaId());
+//    Store store = storeService.getStore(serviceAreaRequestDto.storeId());
+
+    Area area = new Area(serviceAreaRequestDto.areaId());
+    Store store = new Store(serviceAreaRequestDto.storeId());
+
     ServiceAreaResponseDto serviceAreaResponseDto = serviceAreaService.createServiceArea(
         new ServiceAreaCommand(serviceAreaRequestDto, area, store));
 
@@ -49,9 +53,9 @@ public class ServiceAreaController {
   }
 
   // 가게 ID로 배달 가능 지역 조회
-  @GetMapping("/areas/{storeId}")
+  @GetMapping("/areas")
   public ResponseEntity<ResponseDto<List<ServiceAreaResponseDto>>> getAreasByStoreId(
-      @PathVariable UUID storeId) {
+      @RequestParam UUID storeId) {
 
     List<ServiceAreaResponseDto> serviceAreaResponseDtos = serviceAreaService.getAreasByStoreId(
         storeId);
@@ -59,18 +63,24 @@ public class ServiceAreaController {
   }
 
   // 지역 ID로 해당 지역의 가게 조회 -> 페이징 처리 필요
-  @GetMapping("/stores/{areaId}")
+  @GetMapping("/stores")
   public ResponseEntity<ResponseDto<Page<ServiceAreaResponseDto>>> getStoresByAreaId(
-      @PathVariable UUID areaId,
-      @RequestParam(defaultValue = "1") int page,
-      @RequestParam(defaultValue = "10") int size,
-      @RequestParam(defaultValue = "id") String sortBy,
-      @RequestParam(defaultValue = "true") boolean isAsc) {
+      @RequestParam(
+          name = "areaId",
+          required = false,
+          defaultValue = ""
+      ) UUID areaId,
+      @PageableDefault(
+          size = 10,
+          page = 1,
+          sort = {"createdAt", "updatedAt"},
+          direction = Direction.ASC
+      ) Pageable pageable) {
 
-    Pageable pageable = PageableUtils.createPageable(page, size, sortBy, isAsc);
+    Pageable validatedPageable = PageableUtils.validatePageable(pageable);
 
     Page<ServiceAreaResponseDto> serviceAreaResponseDto = serviceAreaService.getStoresByAreaId(
-        areaId, pageable);
+        areaId, validatedPageable);
     return ResponseEntity.ok(ResponseDto.success(serviceAreaResponseDto));
   }
