+++ src/main/java/com/nineteen/omp/store/service/ServiceAreaService.java
@@ -16,68 +16,13 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@Service
-@RequiredArgsConstructor
-public class ServiceAreaService {
+public interface ServiceAreaService {
 
-  private final ServiceAreaRepository serviceAreaRepository;
+  void deleteServiceArea(UUID areaId, UUID storeId);
 
-  @Transactional
-  public ServiceAreaResponseDto createServiceArea(ServiceAreaCommand serviceAreaCommand) {
-    boolean isExists = serviceAreaRepository.existsByAreaIdAndStoreId(
-        serviceAreaCommand.area().getId(),
-        serviceAreaCommand.store().getId());
-    if (isExists) {
-      throw new ServiceAreaException(ServiceAreaExceptionCode.SERVICE_AREA_IS_DUPLICATED);
-    }
+  Page<ServiceAreaResponseDto> getStoresByAreaId(UUID areaId, Pageable validatedPageable);
 
-    ServiceAreaData serviceAreaData = new ServiceAreaData(serviceAreaCommand.area(),
-        serviceAreaCommand.store());
+  List<ServiceAreaResponseDto> getAreasByStoreId(UUID storeId);
 
-    ServiceArea savedServiceArea = serviceAreaRepository.save(serviceAreaData.toEntity());
-
-    return toResponseDto(savedServiceArea);
-  }
-
-  @Transactional(readOnly = true)
-  public List<ServiceAreaResponseDto> getAreasByStoreId(UUID storeId) {
-    boolean isExists = serviceAreaRepository.existsByStoreId(storeId);
-    if (!isExists) {
-      throw new ServiceAreaException(ServiceAreaExceptionCode.SERVICE_AREA_NOT_FOUND);
-    }
-
-    List<ServiceArea> serviceAreas = serviceAreaRepository.findAllByStoreId(storeId);
-
-    if (serviceAreas.isEmpty()) {
-      throw new ServiceAreaException(ServiceAreaExceptionCode.SERVICE_AREA_NOT_FOUND);
-    }
-
-    return serviceAreas.stream()
-        .map(this::toResponseDto)
-        .collect(Collectors.toList());
-  }
-
-  @Transactional(readOnly = true)
-  public Page<ServiceAreaResponseDto> getStoresByAreaId(UUID areaId, Pageable pageable) {
-    return serviceAreaRepository.getStoresByAreaId(areaId, pageable);
-  }
-
-  @Transactional
-  public Boolean deleteServiceArea(UUID areaId, UUID storeId) {
-    ServiceArea serviceArea = serviceAreaRepository.findByAreaIdAndStoreId(areaId, storeId)
-        .orElseThrow(
-            () -> new ServiceAreaException(ServiceAreaExceptionCode.SERVICE_AREA_NOT_FOUND));
-
-    serviceAreaRepository.delete(serviceArea);
-    return true;
-  }
-
-
-  private ServiceAreaResponseDto toResponseDto(ServiceArea serviceArea) {
-    return new ServiceAreaResponseDto(
-        serviceArea.getId(),
-        serviceArea.getArea().getId(),
-        serviceArea.getStore().getId()
-    );
-  }
+  ServiceAreaResponseDto createServiceArea(ServiceAreaCommand serviceAreaCommand);
 }
